add np.datetime64 support to cffi and python views

+++ pygdf/column.py
@@ -175,7 +175,8 @@ class Column(object):
         """
         return _gdf.columnview(size=self._data.size,
                                data=self._data,
-                               mask=self._mask)
+                               mask=self._mask,
+                               dtype=self.dtype)
 
     def set_mask(self, mask, null_count=None):
         """Create new Column by setting the mask
